# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hostreamly Media API",
    "version": "2.0.0",
    "description": "API completa para la plataforma Hostreamly Media con funcionalidades empresariales avanzadas. Esta API proporciona endpoints para gestión de videos, streaming en vivo, analíticas avanzadas, webhooks personalizados, formularios de contacto, enlaces de descarga controlados y mucho más.",
    "termsOfService": "https://hostreamly.com/terms",
    "contact": {
      "name": "Hostreamly Support",
      "email": "support@hostreamly.com",
      "url": "https://hostreamly.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.hostreamly.com/api",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Código de error específico",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del error",
            "example": "Los datos proporcionados no son válidos"
          },
          "details": {
            "type": "array",
            "description": "Detalles adicionales del error",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "El email debe tener un formato válido"
                }
              }
            }
          }
        },
        "example": {
          "error": "VALIDATION_ERROR",
          "message": "Los datos proporcionados no son válidos",
          "details": [
            {
              "field": "email",
              "message": "El email debe tener un formato válido"
            }
          ]
        }
      },
...
      "DownloadLink": {
        "type": "object",
        "required": ["video_id"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador único del enlace de descarga",
            "example": "dl_1234567890abcdef",
            "readOnly": true
          },
          "token": {
            "type": "string",
            "description": "Token único para acceder a la descarga",
            "example": "tok_abcdef1234567890",
            "readOnly": true
          },
          "video_id": {
            "type": "string",
            "description": "ID del video asociado",
            "example": "vid_1234567890"
          },
          "user_id": {
            "type": "string",
            "description": "ID del usuario propietario",
            "example": "usr_1234567890"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de expiración del enlace",
            "example": "2024-01-20T10:30:00Z"
          },
          "max_downloads": {
            "type": "integer",
            "description": "Número máximo de descargas permitidas",
            "example": 5,
            "minimum": 1,
            "default": 1
          },
          "download_count": {
            "type": "integer",
            "description": "Número actual de descargas realizadas",
            "example": 2,
            "readOnly": true
          },
          "ip_restrictions": {
            "type": "array",
            "description": "Lista de IPs o rangos CIDR permitidos",
            "items": {
              "type": "string"
            },
            "example": ["192.168.1.0/24", "10.0.0.1"]
          },
          "require_auth": {
            "type": "boolean",
            "description": "Si requiere autenticación para descargar",
            "example": false,
            "default": false
          },
          "notify_on_download": {
            "type": "boolean",
            "description": "Si enviar notificación al descargar",
            "example": true,
            "default": false
          },
          "status": {
            "type": "string",
            "enum": ["active", "expired", "exhausted", "disabled"],
            "description": "Estado actual del enlace",
            "example": "active"
          }
        },
        "example": {
          "id": "dl_1234567890abcdef",
          "token": "tok_abcdef1234567890",
          "video_id": "vid_1234567890",
          "user_id": "usr_1234567890",
          "expires_at": "2024-01-20T10:30:00Z",
          "max_downloads": 5,
          "download_count": 2,
          "ip_restrictions": ["192.168.1.0/24"],
          "require_auth": false,
          "notify_on_download": true,
          "status": "active"
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/downloads": {
      "post": {
        "summary": "Crear enlace de descarga",
        "description": "Crea un enlace de descarga controlado para un video específico con permisos y configuración de expiración. Útil para compartir videos de forma segura con acceso limitado.",
        "tags": ["Downloads"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["videoId"],
                "properties": {
                  "videoId": {
                    "type": "string",
                    "description": "ID del video para el cual crear el enlace de descarga",
                    "example": "vid_1234567890abcdef"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Fecha y hora de expiración del enlace (ISO 8601)",
                    "example": "2024-01-22T10:30:00Z"
                  },
                  "maxDownloads": {
                    "type": "integer",
                    "description": "Número máximo de descargas permitidas",
                    "minimum": 1,
                    "maximum": 1000,
                    "example": 5
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña opcional para proteger el enlace",
                    "minLength": 4,
                    "maxLength": 50,
                    "example": "mi-password-seguro"
                  },
                  "allowedIPs": {
                    "type": "array",
                    "description": "Lista de direcciones IP permitidas para descargar",
                    "items": {
                      "type": "string",
                      "format": "ipv4"
                    },
                    "example": ["192.168.1.100", "10.0.0.50"]
                  },
                  "notifyOnDownload": {
                    "type": "boolean",
                    "description": "Si enviar notificación cuando se use el enlace",
                    "default": false,
                    "example": true
                  }
                }
              },
              "examples": {
                "simple_download": {
                  "summary": "Enlace de descarga simple",
                  "description": "Enlace básico con expiración de 7 días",
                  "value": {
                    "videoId": "vid_1234567890abcdef",
                    "expiresAt": "2024-01-22T10:30:00Z",
                    "maxDownloads": 10
                  }
                },
                "secure_download": {
                  "summary": "Enlace de descarga seguro",
                  "description": "Enlace protegido con contraseña y restricción de IP",
                  "value": {
                    "videoId": "vid_1234567890abcdef",
                    "expiresAt": "2024-01-16T23:59:59Z",
                    "maxDownloads": 1,
                    "password": "video-confidencial-2024",
                    "allowedIPs": ["192.168.1.100"],
                    "notifyOnDownload": true
                  }
                },
                "team_sharing": {
                  "summary": "Compartir con equipo",
                  "description": "Enlace para compartir con múltiples miembros del equipo",
                  "value": {
                    "videoId": "vid_1234567890abcdef",
                    "expiresAt": "2024-01-30T00:00:00Z",
                    "maxDownloads": 50,
                    "notifyOnDownload": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Enlace de descarga creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DownloadLink"
                    },
                    "message": {
                      "type": "string",
                      "example": "Enlace de descarga creado exitosamente"
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "dl_9876543210fedcba",
                    "token": "dlt_abcdef1234567890",
                    "videoId": "vid_1234567890abcdef",
                    "downloadUrl": "https://api.hostreamly.com/downloads/dlt_abcdef1234567890",
                    "expiresAt": "2024-01-22T10:30:00Z",
                    "maxDownloads": 5,
                    "downloadCount": 0,
                    "isActive": true,
                    "createdAt": "2024-01-15T10:30:00Z"
                  },
                  "message": "Enlace de descarga creado exitosamente"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalid_video": {
                    "summary": "Video no encontrado",
                    "value": {
                      "error": "VALIDATION_ERROR",
                      "message": "El video especificado no existe o no tienes permisos para acceder a él",
                      "details": [
                        {
                          "field": "videoId",
                          "message": "Video no encontrado"
                        }
                      ]
                    }
                  },
                  "invalid_expiration": {
                    "summary": "Fecha de expiración inválida",
                    "value": {
                      "error": "VALIDATION_ERROR",
                      "message": "La fecha de expiración debe ser futura",
                      "details": [
                        {
                          "field": "expiresAt",
                          "message": "La fecha debe ser posterior a la fecha actual"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de tasa excedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "RATE_LIMIT_EXCEEDED",
                  "message": "Has excedido el límite de creación de enlaces de descarga. Intenta nuevamente en 1 hora."
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar enlaces de descarga",
        "description": "Obtiene una lista de todos los enlaces de descarga creados por el usuario autenticado",
        "tags": ["Downloads"],
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "Filtrar por ID de video específico",
            "required": false,
            "schema": {
              "type": "string",
              "example": "vid_1234567890abcdef"
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "Filtrar por enlaces activos o inactivos",
            "required": false,
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número máximo de resultados a devolver",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Número de resultados a omitir (para paginación)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de enlaces de descarga recuperada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DownloadLink"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 45
                        },
                        "limit": {
                          "type": "integer",
                          "example": 20
                        },
                        "offset": {
                          "type": "integer",
                          "example": 0
                        },
                        "hasMore": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/contact": {
      "post": {
        "summary": "Crear formulario de contacto",
        "description": "Crea un nuevo formulario de contacto personalizable con capacidad de upload de video integrado",
        "tags": ["Contact Forms"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookConfig"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}